Fixed SetAirdrop
