+++ Models/AirdropSpawn.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Xml.Serialization;
+
+namespace RestoreMonarchy.AirdropManager.Models
+{
+    public class AirdropSpawn
+    {
+        [XmlAttribute("AirdropId")]
+        public ushort AirdropId { get; set; }
+        public Position Position { get; set; }
+    }
+}
